btc data in own file
